requirements.txt
alembic==1.13.2
Authlib==1.3.1
blinker==1.7.0
cachelib==0.13.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
flask-mongoengine==1.0.0
Flask-OAuthlib==0.9.6
flask-oidc==2.2.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
mongoengine==0.28.2
oauthlib==2.1.0
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
pymongo==4.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.2
Werkzeug==3.0.1
WTForms==3.1.2
